Switch to programWithFlags for init
So the correct way to do that initialization is to pass the init data in
the embed call as a flag, which means we use Html.programWithFlags
instead of .program.  This means we're initialized from the start, so
we're never waiting for the data and the model gets simpler becasue we
can lose the maybe.