Refactor to reduce coupling
Move the model.value /= model'.value test for field
canges back into Field.elm -- which feels better encapslated

Add update' to Field which does update and retuns Maybe Msg
which is None if there is no action to perform, and Just Saved
if we want to get back a Saved Msg. The neat thing here
is that the parent/client module no longer needs to know
about the Field.Msg values. It just gets passedone back
by update' and tunrs it into a Cmd mapping over it
using Cmd.None as the defalt.