WIP (Untested) Change how Result.id is optional for placeholder
Switch from explicit modelling of placeholder without id and result
objects with id to a model where id is optional (maybe). This makes it
cleaner to keep responsibility for forming the http request to save a
result inside Result module, removes all the ugly pattern matches for
(id, result) from many functionss and removes some other complexity I
Wasn't expecting.

Now I'm too tired to test this tonight but I'm commiting it anyway so I
can go to bed. But I think this covers the step to include the
placeholder in the model. Next up, if this hasnt broken anything, render
it in the view.

Oh for some automatred testing. At the moment I do all my testing by
hand and the most obvious way to automate would be with some funcitonal
tests, e.g. with SPlinter and python. What is the Elm way to test
this sort of stuff at high level?